requirements.txt
requests
pandas
python-dotenv
supabase
Flask
python-telegram-bot
